WIP fixed collec drop

+++ src/services/db/db.js
@@ -42,10 +42,11 @@ class quizDB {
     * @param collection String, collection to drop, Required
     */
     async dropCollection(collection) {
-        let collectionsCursor = await this.db.listCollections();
-        if( (await collectionsCursor.toArray()).indexOf(collection) >= 0 ){
-            await this.db.dropCollection(collection);
-        }  
+        let collectionsCursor = await this.db.listCollections({}, { nameOnly: true });
+        let collections = await collectionsCursor.toArray();
+        let collectionNames = collections.map(c =>c.name);
+
+        if (collectionNames.indexOf(collection) >= 0) await this.db.dropCollection(collection);
     }
 
     /**
@@ -55,9 +56,9 @@ class quizDB {
     */
     async createCollection(collection) {
         let collectionsCursor = await this.db.listCollections();
-        if( (await collectionsCursor.toArray()).indexOf(collection) < 0 ){
+        if ((await collectionsCursor.toArray()).indexOf(collection) < 0) {
             await this.db.createCollection(collection);
-        } 
+        }
     }
 
     /**
